Pass firewall ID to created tokens
The firewall ID must be passed to the UsernamePasswordToken (or its sub classes) to ensure that authentication is done correctly if there are multiple firewalls.
Having a hard-coded value prevents the AuthProvider to be used for more than one firewall.

In addition, a provider key which does not match the actual firewall ID prevents the Twig functions logout_path() and logout_url() to work correctly.